Remove single escaped Greek character
So far, Zotero is replacing fields with 1 one match, but not ≥1 match,
despite the fact that I have the global `g` switch in my regex. If there
is ≥1 match, Zotero doesn't replace any of the matches.

Also note, in the `items` object I'm creating for debugging purposes
prior to actually replacing any metadata, Zotero is not serializing the
object properly, so there's no comma after the title field.