Add notes on file renaming with Zotero 7
Zotero 7 includes many improvements, but may not be backwards compatible
with Zotero 6 in some ways. Some plugins have not been updated for
Zotero 7 and will stop working after updating.

One such plugin is ZotFile. Zotfile was useful for automatically moving
and renaming files when they were added to a Zotero library. The ZotFile
maintainer is not planning to update ZotFile for Zotero 7. ZotMoov is an
alternative to ZotFile that is compatible with Zotero 7.

This commit will update the Zotero guide in this repo for ZotMoov and
the related changes needed for Zotero 7. The transition from ZotFile to
ZotMoov is straightforward for the most part. The major difference is
that ZotMoov does not have its own file renaming rules like ZotFile did.
However, Zotero 7 does include its own file renaming features that can
be used instead. An example file renaming template will be included that
is close to matching the previous ZotFile behavior. The main difference
is in title truncation behavior. The behavior of the Zotero 7 `truncate`
field is slightly different from how ZotFile truncated titles. Zotero
will truncate at the given limit, regardless of the final character in
the title. ZotFile truncated to the last word under the character limit.

For reference, the file renaming template currently looks like this:

```
{{ if {{ authors }} == {{ authors max="3" }} }}
{{ authors case="snake" join="_" name="family" suffix="_" }}
{{ else }}
{{ authors case="snake" join="_" max="1" name="family" suffix="_et_al_" }}
{{ endif }}
{{ year suffix="_" }}
{{ title case="snake" replaceFrom="(.*)([\.:?!]| - )(.*)" replaceTo="$1" truncate="80" }}
```

Related:

https://www.zotero.org/blog/zotero-7/
https://www.zotero.org/support/file_renaming
https://github.com/jlegewie/zotfile/issues/655
https://github.com/wileyyugioh/zotmoov

Source code for ZotFile title truncation behavior:
https://github.com/jlegewie/zotfile/blob/7b2c98b5a08ddfcfdcc99f6d3fa1a8aff776eb01/chrome/content/zotfile/wildcards.js#L79-L83